Add parameter '-D' for alternative APKINDEX description
This patch allows to set a nice description for the APKINDEX, in case
the aport that is being built is not inside a git repository.

I have tested it, and it behaves exactly like without the patch,
even when git is not installed, or the folder is not inside a git repository:
The `|| true` at the end of the DESCRIPTION= line makes sure, aport
does not get aborted, just like it does not get aborted in that case
as of now, as the `git describe` command gets executed in a subshell.