Don't print 'git: not found' errors
abuild, as packaged in Alpine Linux, does not depend on git. But when
you use it without git, it will print out errors like the following:

/usr/bin/abuild: line 2554: git: not found

With this commit, it saves the git_path in the beginning (just like
abuild_path). Later in the code it does not try to run git if that
variable is empty.

Notably `abuild rootbld` is already checking whether `abuild-rootbld`
is installed, and that subpackage of `abuild` does already depend on
`git`. So no additional check was added before using `git ` inside
`rootbld`.