feat: add amove
Amove is a utility function for subpackage splitting.
It takes up to 2 arguments:
$1 is a relative directory (or file) (/ is dropped if present)
$2 is a regex-like format, as taken by find(1p)/-name

If only $1 is specified, it will move $1 from pkgdir to subpkgdir,
creating what is needed and cleaning up after itself appropriately.
If $2 is specified, $1 should be a directory (not checked - unneeded).
When $2 is specified and $1 is a directory, all files directly inside of
$1 that match the $2 namespec will be copied instead.