tests: set APORTSDIR for abuild tests
If `APORTSDIR` is not set, abuild tries to find it. One way it does this
is by trying to look at the url for the git remote 'origin'. This fails
either if there is no git repository, or the git repository has no
remote called 'origin'.

Instead of having abuild guess the location, provide it by exporting
APORTSDIR.