abuild: Do not enable nftables ruleset automatically
As discussed in the aports merge request, we think it is better to not
enable the ruleset by default. Mainly because then an installed package
would make holes in the firewall, even if the users does not run the
software.

If we install the ruleset in a different folder, the users can symlink
what they needs. They can symlink all hooks, or the directory, or use
them however they wants.

This store the nft ruleset in /usr/share/nftables.avail/ as discussed
upstream:

https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/67137#note_458749