abuild: improve man page compression
Since the sort(1) was added (for reproducible builds) the inodes were
handled in order. Meaning multiple identical inodes (aka hardlinks)
were handled in consecutive iterations. By remembering the previous
inode/name hardlinks can be handled with only a single find(1) call.

This saves around a minute when compressing the openssl man pages.

Since the stat(1) call doesn't follow symlinks (by default) it was also
possible to merge symlink handling into it.

Also fixes escaping of -name arguments and use -r for best practices.

There previously also seemed to be a bug since islink=1 was only ever
set in a subshell which made it a noop. Not sure if this had any effect
though.