abuild: add udev splitfunc
Many packages provide udev rules, which enhance their functionality or
simplify their setup on systems with udev. Some of these packages split
the rules into a $pkgname-udev subpackage.

Ideally, all packages should split these rules into a $pkgname-udev
subpackage to avoid installing rules in situations where they are
unnecessary and unused. After updating around a dozen packages to
include an -udev subpackage, it became clear that this introduced a
great deal of repetition, given that the udev() function was identical
in all of them.

This change implement support for a -udev subpackage, following the same
approach as -openrc and -systemd subpackages. Packages which provide
udev rules simply need to include `subpackages="… $pkgname-dev"`, and
the rules themselves will automatically be moved into the subpackage.