abump: implement -r to build using rootbld
Allow using `abump -r` to build using rootbld instead of building on the
host. This is especially useful when a build dependency conflicts with a
package installed on the host.

Using `abump -r` also detects new dependencies which are present in the
host but not explicitly declared in the APKBUILD.